add more tests and move date time tests to their own class
